requirements.txt
streamlit
llama-cpp-python
pandas
google-cloud-storage